Day15/requirements.txt
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
groq>=0.1.0
python-dotenv>=1.0.0
